Kill all the old stuff, and replace with the new.
